Add files via upload
Added timer and camera (plus picture taking) features. In order words, finished my and Khoula's part. But it still displays the image it takes. This needs to be removed, but only once everyone elses part is done. 

Recall, the folder in the main that actually has all the proper files is called: rockpaperscissors_appp.